refactor(admin): push AdminInfraController logic into focused services (max-params 2/3)
AdminInfraController injected 7 deps (adminInfra, embedder, intent,
changefeed, activity, throttler, config). Move the per-component health
grid into HealthComponentsService (embedder, intent, changefeed; the
DB-liveness result is passed in, config flags read via process.env) and
the live operational snapshots into LiveSnapshotService (throttler,
activity). The controller keeps HTTP plumbing only (adminInfra,
healthComponents, liveSnapshot), 3 deps.

Named-dep test helper + the now/throttler/health-components contract
specs updated to the new shape.

Part of the max-params=3 god-class split program.
