+++ src/admin/admin.module.ts
@@ -16,6 +16,8 @@ import { AdminJobsController } from './admin-jobs.controller';
 import { AdminOpsController } from './admin-ops.controller';
 import { AdminInfraController } from './admin-infra.controller';
 import { AdminInfraService } from './admin-infra.service';
+import { HealthComponentsService } from './health-components.service';
+import { LiveSnapshotService } from './live-snapshot.service';
 import { AdminService } from './admin.service';
 import { OperatorActionService } from './operator-action.service';
 import { OperatorActionInterceptor } from './operator-action.interceptor';
@@ -56,6 +58,8 @@ import { ConfigInspectorService } from './config-inspector.service';
   providers: [
     AdminService,
     AdminInfraService,
+    HealthComponentsService,
+    LiveSnapshotService,
     ScenarioRunnerService,
     BaselineService,
     ChatRouterCacheService,
